[hotfix][runtime] Deduplicate buffersInBacklog code in Pipelined and Spillable subtartitions
